requirements.txt
tensorflow==2.13.0
opencv-python
numpy
matplotlib
scikit-learn
seaborn
fastapi
uvicorn
python-multipart
streamlit
pillow
